+++ sparql_service/src/graph_collection.rs
@@ -1,4 +1,4 @@
-use crate::GraphDescription;
+use crate::NamedGraphDescription;
 use serde::{Deserialize, Serialize};
 use srdf::IriOrBlankNode;
 use std::{collections::HashSet, fmt::Display, hash::Hash};
@@ -6,8 +6,9 @@ use std::{collections::HashSet, fmt::Display, hash::Hash};
 #[derive(Clone, PartialEq, Eq, Debug, Serialize, Deserialize)]
 pub struct GraphCollection {
     id: IriOrBlankNode,
+
     #[serde(skip_serializing_if = "HashSet::is_empty")]
-    collection: HashSet<GraphDescription>,
+    collection: HashSet<NamedGraphDescription>,
 }
 
 impl GraphCollection {
@@ -16,6 +17,11 @@ impl GraphCollection {
             id: id.clone(),
             collection: HashSet::new(),
         }
+    }
+
+    pub fn with_collection<I: Iterator<Item = NamedGraphDescription>>(mut self, graphs: I) -> Self {
+        self.collection = HashSet::from_iter(graphs);
+        self
     }
 }
 
@@ -27,6 +33,10 @@ impl Hash for GraphCollection {
 
 impl Display for GraphCollection {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "Id: {}", self.id)
+        write!(f, "Id: {}", self.id)?;
+        for graph in &self.collection {
+            writeln!(f, "\nGraph: {}", graph)?;
+        }
+        Ok(())
     }
 }